refactor(web): extract reusable components and enforce separation of concerns
- Extract AddChildForm into shared component, eliminating duplication
  between child-selector.tsx and dashboard/index.tsx (~115 lines deduped)
- Extract JournalCard and JournalForm from journal route into
  components/journal/ for reusability and testability
- Extract SymptomCard + dimensionLabels from symptoms route into
  components/symptoms/symptom-card.tsx
- Route files become pure orchestrators: composition and layout only
- No behavior changes — pure structural refactoring
